requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
aiohttp>=3.9.0
pypdf2>=3.0.0
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
pdfplumber>=0.10.0
python-multipart>=0.0.6

